(size_t) cast is already part of Memcpy() macro definition
git-svn-id: https://svn.r-project.org/R/trunk@46147 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/lapack/Lapack.c
@@ -82,7 +82,7 @@ static SEXP modLa_svd(SEXP jobu, SEXP jobv, SEXP x, SEXP s, SEXP u, SEXP v, SEXP
     p = xdims[1];
     xvals = (double *)R_alloc(n * p, sizeof(double));
     /* work on a copy of x */
-    Memcpy(xvals, REAL(x), (size_t)(n * p));
+    Memcpy(xvals, REAL(x), n * p);
 
     {
         int ldu = INTEGER(getAttrib(u, R_DimSymbol))[0], ldvt = INTEGER(getAttrib(v, R_DimSymbol))[0];
@@ -237,7 +237,7 @@ static SEXP modLa_rg(SEXP x, SEXP only_values)
 
     xvals = (double *)R_alloc(n * n, sizeof(double));
     /* work on a copy of x */
-    Memcpy(xvals, REAL(x), (size_t)(n * n));
+    Memcpy(xvals, REAL(x), n * n);
     ov = asLogical(only_values);
     if (ov == NA_LOGICAL)
         error(_("invalid '%s' argument"), "only.values");
@@ -426,7 +426,7 @@ static SEXP modLa_zgecon(SEXP A, SEXP norm)
     /* Compute the LU-decomposition and overwrite 'x' with result;
      * working on a copy of A : */
     avals = (Rcomplex *)R_alloc(n * n, sizeof(Rcomplex));
-    Memcpy(avals, COMPLEX(A), (size_t)(n * n));
+    Memcpy(avals, COMPLEX(A), n * n);
     F77_CALL(zgetrf)
     (&n, &n, avals, &n,
      /* iwork: */ (int *)R_alloc(n, sizeof(int)), &info);
@@ -513,7 +513,7 @@ static SEXP modLa_zgesv(SEXP A, SEXP Bin)
 
     avals = (Rcomplex *)R_alloc(n * n, sizeof(Rcomplex));
     /* work on a copy of x */
-    Memcpy(avals, COMPLEX(A), (size_t)(n * n));
+    Memcpy(avals, COMPLEX(A), n * n);
     F77_CALL(zgesv)(&n, &p, avals, &n, ipiv, COMPLEX(B), &n, &info);
     if (info < 0)
         error(_("argument %d of Lapack routine %s had invalid value"), -info, "zgesv");
@@ -776,7 +776,7 @@ static SEXP modLa_rg_cmplx(SEXP x, SEXP only_values)
 
     xvals = (Rcomplex *)R_alloc(n * n, sizeof(Rcomplex));
     /* work on a copy of x */
-    Memcpy(xvals, COMPLEX(x), (size_t)(n * n));
+    Memcpy(xvals, COMPLEX(x), n * n);
     ov = asLogical(only_values);
     if (ov == NA_LOGICAL)
         error(_("invalid '%s' argument"), "only.values");
@@ -936,7 +936,7 @@ static SEXP modLa_dgesv(SEXP A, SEXP Bin, SEXP tolin)
 
     avals = (double *)R_alloc(n * n, sizeof(double));
     /* work on a copy of A */
-    Memcpy(avals, REAL(A), (size_t)(n * n));
+    Memcpy(avals, REAL(A), n * n);
     F77_CALL(dgesv)(&n, &p, avals, &n, ipiv, REAL(B), &n, &info);
     if (info < 0)
         error(_("argument %d of Lapack routine %s had invalid value"), -info, "dgesv");
